requirements.txt
torch
streamlit
matplotlib
numpy
Pillow
